Escape table name, only when it's not * [GH-1060]
In my previous commit 937da5c I had removed the table name from being
escape altogether because a value of '*' would generate an incorrect
grant statement. As pointed out in ticket [GH-1060] if you specify a
table name this would break the grant generation.

Now the table name is escaped only if it's not '*' which allows both
use cases to work.