Reload renderers for local pillar after gitfs (#32912)
Because of the way the renderers are lazy loaded when using ext_pillar_first
and gitfs external pillar, the renderers for the local pillar object were
being set to the last loaded gitfs pillar.

This was causing the local pillar to fail to render.  This small fix
forces a reload of the renderers after all the external pillars are done
loading.